+++ index.ts
@@ -5,6 +5,15 @@ interface XXssProtectionOptions {
   setOnOldIE?: boolean;
 }
 
+function doesUserAgentMatchOldInternetExplorer(userAgent: string | undefined): boolean {
+  if (!userAgent) {
+    return false;
+  }
+
+  const matches = /msie\s*(\d{1,2})/i.exec(userAgent);
+  return matches ? parseFloat(matches[1]) < 9 : false;
+}
+
 export = function xXssProtection (options: XXssProtectionOptions = {}) {
   let headerValue = '1; mode=block';
   if (options.reportUri) {
@@ -18,15 +27,7 @@ export = function xXssProtection (options: XXssProtectionOptions = {}) {
     };
   } else {
     return function xXssProtection (req: IncomingMessage, res: ServerResponse, next: () => void) {
-      const matches = /msie\s*(\d+)/i.exec(req.headers['user-agent'] || '');
-
-      let value;
-      if (!matches || parseFloat(matches[1]) >= 9) {
-        value = headerValue;
-      } else {
-        value = '0';
-      }
-
+      const value = doesUserAgentMatchOldInternetExplorer(req.headers['user-agent']) ? '0' : headerValue;
       res.setHeader('X-XSS-Protection', value);
       next();
     };